Move OrtValue binding to a separate file Move IOBinding into separate file.

+++ onnxruntime/test/python/onnxruntime_test_python_iobinding.py
@@ -25,7 +25,7 @@ class TestIOBinding(unittest.TestCase):
         return np.array([[2.0, 8.0], [18.0, 32.0], [50.0, 72.0]], dtype=np.float32)
 
     def test_ort_value_properties(self):
-        ort_value = onnxruntime.OrtValue.ortvalue_from_numpy(np.array([2.0, 4.0]), 'cpu', 0)
+        ort_value = onnxruntime.OrtValue.ortvalue_from_numpy(np.array([2.0, 4.0], dtype=np.float32), 'cpu', 0)
         self.assertTrue(ort_value.is_tensor())
         shape = ort_value.shape()
         self.assertTrue([2], shape)
